+++ CHANGELOG.md
@@ -4,5 +4,6 @@
 * [New] The class `ResourceWatcher` uses content hash instead of dates to detect changes.
 * [New] The method `findChanges` from the class `ResourceWatcher` returns an object type `ResourceWatcherResult` with all the information about files changes.
 * [New] Added a new method `initialize` to the class  `ResourceWatcher`. This method warm up the cache just in case.
+* [Improved] The class `ResourceCacheFile` has been renamed to `ResourceCachePhpFile` to improve the readability.
 * [Delete] Deleted the method `isSearching` from the class `ResourceWatcher`.
 * [Delete] Deleted the method `setFinder` from the class `ResourceWatcher`. Now, the finder is passed as constructor argument.